+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Performance increase of finding neighbors in GraphSpace with r > 1.
 - New wrapping function `nearby_agents` that returns an iterable of neighboring agents.
 - New public `schedule` function for writing custom loops.
+- Mixed models are supported in data collection methods.
 - `random_agent(model, condition)` allows obtaining random agents that satisfy given condition.
 
 ## Breaking changes